requirements.txt
streamlit
Pillow
requests
duckduckgo-search
opencv-python-headless
numpy